sht: change implicit cast to explicit.
Fixes a warning on 32-bit RasPi.

+++ attic/sht.c
@@ -90,7 +90,7 @@ main (
 		exit (0);
 	}
 
-	unit = strtoul(argv[1], &argp, 10);
+	unit = (int)strtoul(argv[1], &argp, 10);
 	if (argp == argv[1])
 		unit = 2;
 	else if (*argp == ':')
